S11P12C202-123 Add : 성적리스트 조회 시 상세내역 추가

See merge request s11-webmobile1-sub2/S11P12C202!22

+++ backend/src/main/java/com/studycow/dto/ScoreDetailDto.java
@@ -10,8 +10,8 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ScoreDetailDto {
-    private int wrongDetailId;
-    private int scoreId;
+    private Long wrongDetailId;
+    private Long scoreId;
     private int catCode;
     private String catName;
     private int wrongCnt;
